First implementation of ChaseStrategy (the algorithm responsible for determine the next direction that has to take the ghost). Pacman is now a Singleton, the reason for this is that the Ghosts need to know always where the Pacman is. And we won't have more than one Pacman.

+++ TP1/src/main/java/com/grupo2/character/Coordinate.java
@@ -58,10 +58,10 @@ public class Coordinate {
 	 * @param otherCoordinate Coordinate to compute distance.
 	 * @return distance
 	 */
-	public float distanceTo(Coordinate otherCoordinate) {
+	public int distanceTo(Coordinate otherCoordinate) {
 		int distanceX = this.getX() - otherCoordinate.getX();
 		int distanceY = this.getY() - otherCoordinate.getY();
 		int valuedDistance = distanceX * distanceX + distanceY * distanceY;
-		return (float) sqrt(valuedDistance);
+		return (int) sqrt(valuedDistance);
 	}
 }
